+++ ft_calloc.c
@@ -6,13 +6,13 @@
 /*   By: lburnet <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/11/25 13:54:40 by lburnet           #+#    #+#             */
-/*   Updated: 2020/11/25 14:38:17 by lburnet          ###   ########lyon.fr   */
+/*   Updated: 2020/11/27 10:30:09 by lburnet          ###   ########lyon.fr   */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void	*ft_calloc(unsigned int count, unsigned int size)
+void	*ft_calloc(size_t count, size_t size)
 {
 	void	*p;
