+++ ft_memchr.c
@@ -6,13 +6,13 @@
 /*   By: lburnet <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/11/24 11:15:53 by lburnet           #+#    #+#             */
-/*   Updated: 2020/11/24 15:21:58 by lburnet          ###   ########lyon.fr   */
+/*   Updated: 2020/11/27 10:15:20 by lburnet          ###   ########lyon.fr   */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void	*ft_memchr(const void *s, int c, unsigned int n)
+void	*ft_memchr(const void *s, int c, size_t n)
 {
 	unsigned char *p;
