+++ ft_strdup.c
@@ -6,7 +6,7 @@
 /*   By: lburnet <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/11/25 14:39:55 by lburnet           #+#    #+#             */
-/*   Updated: 2020/11/25 15:56:40 by lburnet          ###   ########lyon.fr   */
+/*   Updated: 2020/11/27 10:31:37 by lburnet          ###   ########lyon.fr   */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -28,8 +28,8 @@ char	*ft_strcpy(char *dest, const char *src)
 
 char	*ft_strdup(const char *src)
 {
-	char			*str;
-	unsigned int	len_src;
+	char		*str;
+	size_t		len_src;
 
 	len_src = ft_strlen(src) + 1;
 	if (!(str = (char *)malloc(len_src * sizeof(char))))
